Adding and modifying structs for lane and road parsing
